請看詳細
2019/2/10 修復
1.修復炎魔前制任務不能進入問題
2.修復紅包蒐集活動-無法累積問題
3.修復道具合成-深淵與凡雷恩武器無法對話問題
4.修復裝備墳墓無法贖回裝備問題
5.修復部分技能技能點滿後仍是0的問題
6.修復女皇無限回血問題
7.新增專業技術村製作道具失敗提醒
8.現在混沌頻道的楓幣獲得提高
9.修正傷害計算(修復這項目前不會有人有感覺)
10.修復冷卻時間
11.修復點裝無法擺入商城問題(目前的裝備如無法擺入 請使用道具欄重整)
12.增加幸運骰子持續時間

+++ src/scripting/NPCConversationManager.java
@@ -2703,11 +2703,11 @@ public class NPCConversationManager extends AbstractPlayerInteraction {
                 eq.setDurability(rs.getInt("durability"));
 
                 eq.setEnhance(rs.getByte("enhance"));
-                eq.setPotential1(rs.getShort("potential1"));
-                eq.setPotential2(rs.getShort("potential2"));
-                eq.setPotential3(rs.getShort("potential3"));
-                eq.setPotential4(rs.getShort("potential4"));
-                eq.setPotential5(rs.getShort("potential5"));
+                eq.setPotential1(rs.getInt("potential1"));
+                eq.setPotential2(rs.getInt("potential2"));
+                eq.setPotential3(rs.getInt("potential3"));
+                eq.setPotential4(rs.getInt("potential4"));
+                eq.setPotential5(rs.getInt("potential5"));
 
 
                 eq.setOwner(rs.getString("owner"));
